update filename update sdrmsd: * add new types: AutomorphismRMSD, Superpose3D * add functions from main as methods to sdrmsd class in favor of organization
