+++ rdock-utils/rdock_utils/sdrmsd/sdrmsd.py
@@ -11,7 +11,8 @@ logger = logging.getLogger("sdrmsd")
 
 Coordinate = tuple[float, float, float]
 SingularValueDecomposition = tuple[numpy.ndarray[float], numpy.ndarray[float], numpy.ndarray[float]]
-RMSDResult = float | tuple[float, numpy.ndarray[float]]
+AutomorphismRMSD = float | tuple[float, numpy.ndarray[float]]
+Superpose3D = tuple[numpy.ndarray[float], float, numpy.ndarray[float]] | tuple[numpy.ndarray[float], float]
 
 
 @dataclass
@@ -19,7 +20,7 @@ class SDRMSDData:
     skipped: list[int] = field(default_factory=list)
     molecules_dict: dict[int, pybel.Molecule] = field(default_factory=dict)  # Save all poses with their dockid
     population: dict[int, int] = field(default_factory=dict)  # Poses to be written
-    out_dict: dict[int, tuple[pybel.Molecule, RMSDResult]] = field(default_factory=dict)
+    out_dict: dict[int, tuple[pybel.Molecule, AutomorphismRMSD]] = field(default_factory=dict)
 
 
 @dataclass
@@ -52,7 +53,7 @@ class SDRMSD:
         reference_mask: numpy.ndarray[bool] | None = None,
         target_mask: numpy.ndarray[bool] | None = None,
         return_rotation_matrix: bool = False,
-    ) -> tuple[numpy.ndarray[float], float, numpy.ndarray[float]] | tuple[numpy.ndarray[float], float]:
+    ) -> Superpose3D:
         """superpose_3D performs 3d superposition using a weighted Kabsch algorithm : http://dx.doi.org/10.1107%2FS0567739476001873 & doi: 10.1529/biophysj.105.066654
         definition : superpose3D(reference, target, weights,refmask,target_mask)
         @parameter 1 :  reference - xyz coordinates of the reference structure (the ligand for instance)
@@ -121,7 +122,7 @@ class SDRMSD:
         # print U
         return (new_coords, rmsd, U) if return_rotation_matrix else (new_coords, rmsd)
 
-    def get_automorphism_rmsd(self, target: pybel.Molecule, molecule: pybel.Molecule) -> RMSDResult:
+    def get_automorphism_rmsd(self, target: pybel.Molecule, molecule: pybel.Molecule) -> AutomorphismRMSD:
         """
         Use Automorphism to reorder target coordinates to match reference coordinates atom order
         for correct RMSD comparison. Only the lowest RMSD will be returned.
@@ -181,8 +182,9 @@ class SDRMSD:
             svd_result = self._handle_svd_linalg_error(reference_tmp, target_tmp)
         return svd_result
 
+    @staticmethod
     def _handle_svd_linalg_error(
-        self, reference_tmp: numpy.ndarray[float], target_tmp: numpy.ndarray[float]
+        reference_tmp: numpy.ndarray[float], target_tmp: numpy.ndarray[float]
     ) -> SingularValueDecomposition | None:
         try:
             dot_product = numpy.dot(numpy.transpose(reference_tmp), target_tmp)
@@ -224,12 +226,101 @@ class SDRMSD:
         for i, atom in enumerate(obmol):
             atom.OBAtom.SetVector(*new_coordinates[i])
 
-    def get_best_matching_pose(self, pose_match_data: PoseMatchData, threshold: float) -> tuple[int | None, float]:
+    def get_best_matching_pose(self, pose_match_data: PoseMatchData) -> tuple[int | None, float]:
         docked_pose = pose_match_data.docked_pose
         molecules_dict = pose_match_data.sdrmsd_data.molecules_dict
         get_rmsd = functools.partial(self.get_automorphism_rmsd, target=docked_pose)
 
         poses_rmsd = ((index, get_rmsd(molecule)) for index, molecule in molecules_dict.items())
-        filtered_by_threshold = (t for t in poses_rmsd if t[1] < threshold)
+        filtered_by_threshold = (t for t in poses_rmsd if t[1] < self.threshold)
 
         return min(filtered_by_threshold, key=lambda t: t[1], default=(None, math.inf))
+
+    def process_and_save_selected_molecules(self, output_sdf: pybel.Outputfile, data: SDRMSDData):
+        for i in sorted(data.out_dict.keys()):
+            molecule, rmsd_result = data.out_dict[i]
+            # Get the number of matches in the thresholding operation
+            population_value = data.population.get(i, 1)
+            self.save_molecule_with_rmsd(output_sdf, molecule, rmsd_result, population_value)
+
+    def save_molecule_with_rmsd(
+        self, output_sdf: pybel.Outputfile, molecule: pybel.Molecule, rmsd: float, population_value: int
+    ) -> None:
+        new_data = pybel.ob.OBPairData()
+        new_data.SetAttribute("RMSD")
+        new_data.SetValue(f"{rmsd:.3f}")
+
+        if self.threshold:
+            pop_data = pybel.ob.OBPairData()
+            pop_data.SetAttribute("Population")
+            pop_data.SetValue(f"{population_value}")
+            molecule.OBMol.CloneData(pop_data)
+
+        molecule.OBMol.CloneData(new_data)  # Add new data
+        output_sdf.write(molecule)
+
+    def get_crystal_pose(self) -> pybel.Molecule:
+        """
+        Read crystal pose file and remove hydrogen atoms. Returns crystal pose molecule.
+        """
+        crystal_pose = next(pybel.readfile("sdf", self.reference_sdf))
+        crystal_pose.removeh()
+        return crystal_pose
+
+    def display_fit_message(self) -> None:
+        message = "FIT" if self.fit else "NOFIT"
+        print(f"POSE\tRMSD_{message}")
+
+    def process_docked_pose(self, docked_pose: pybel.Molecule) -> int:
+        """
+        Remove hydrogen atoms and return the number of atoms in docked pose molecule.
+        """
+        docked_pose.removeh()
+        return len(docked_pose.atoms)
+
+    def calculate_rmsd(self, crystal: pybel.Molecule, docked_pose: pybel.Molecule) -> AutomorphismRMSD:
+        """
+        Perform RMSD calculations and update coordinates if required.
+        """
+        if self.fit:
+            rmsd_result, fitted_result = self.get_automorphism_rmsd(crystal, docked_pose)
+            self.update_coordinates(docked_pose, fitted_result)
+        else:
+            rmsd_result = self.get_automorphism_rmsd(crystal, docked_pose)
+
+        return rmsd_result
+
+    def handle_pose_matching(self, rmsd_result: AutomorphismRMSD, pose_match_data: PoseMatchData) -> None:
+        """
+        Function to handle pose matching and filtering based on 'threshold' parser argument.
+        """
+        if self.threshold:
+            match_pose, best_match_value = self.get_best_matching_pose(pose_match_data)
+            if match_pose is not None:
+                self.print_matching_info(pose_match_data, match_pose, best_match_value)
+            else:
+                self.save_and_print_info(rmsd_result, pose_match_data)
+        else:
+            self.save_and_print_info(rmsd_result, pose_match_data)
+
+    def print_matching_info(self, pose_match_data: PoseMatchData, match_pose: int, best_match_value: float) -> None:
+        index = pose_match_data.pose_index
+        population = pose_match_data.sdrmsd_data.population
+        logger.info(f"Pose {index} matches pose {match_pose} with {best_match_value:.3f} RMSD")
+        population[match_pose] += 1
+
+    def save_and_print_info(self, rmsd_result: AutomorphismRMSD, pose_match_data: PoseMatchData) -> None:
+        """
+        Function to save and print information based on 'out' parser argument.
+        """
+        index = pose_match_data.pose_index
+        docked_pose = pose_match_data.docked_pose
+        out_dict = pose_match_data.sdrmsd_data.out_dict
+        molecules_dict = pose_match_data.sdrmsd_data.molecules_dict
+        population = pose_match_data.sdrmsd_data.population
+
+        if self.out:
+            out_dict[index] = (docked_pose, rmsd_result)
+        print(f"{index}\t{rmsd_result:.2f}")
+        molecules_dict[index] = docked_pose
+        population[index] = 1
